saving to db

+++ src/main/java/com/bierbrauer/gametime/repositories/adventureMessage/AdventureMessageRepo.java
@@ -1,12 +1,12 @@
 package com.bierbrauer.gametime.repositories.adventureMessage;
 
-import com.bierbrauer.gametime.models.AdventureMessage;
+import com.bierbrauer.gametime.models.AdventureMessage.AdventureMessage;
 
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface AdventureMessageRepo extends CrudRepository<AdventureMessage, Long> {
+public interface AdventureMessageRepo extends CrudRepository<AdventureMessage, Integer> {
 
-    Optional<AdventureMessage> findById(Long id);
+    Optional<AdventureMessage> findById(Integer id);
 }
